Some refactoring of submodules
