+++ src/GitTools.jl
@@ -7,7 +7,6 @@ using SHA
 import Base: SHA1
 import LibGit2
 using Printf
-export set_readonly
 
 Base.@kwdef mutable struct MiniProgressBar
     max::Float64 = 1.0
@@ -276,18 +275,5 @@ function tree_hash(root::AbstractString, HashType = SHA.SHA1_CTX)
     return SHA.digest!(ctx)
 end
 
-function set_readonly(path)
-    for (root, dirs, files) in walkdir(path)
-        for file in files
-            filepath = joinpath(root, file)
-            fmode = filemode(filepath)
-            try
-                chmod(filepath, fmode & (typemax(fmode) ⊻ 0o222))
-            catch
-            end
-        end
-    end
-    return nothing
-end
 
 end # module
