+++ src/utils.jl
@@ -1,0 +1,14 @@
+function set_readonly(path)
+    for (root, dirs, files) in walkdir(path)
+        for file in files
+            filepath = joinpath(root, file)
+            fmode = filemode(filepath)
+            try
+                chmod(filepath, fmode & (typemax(fmode) ⊻ 0o222))
+            catch
+            end
+        end
+    end
+    return nothing
+end
+
